requirements.txt
Django==1.10
django-cors-headers==1.1.0
django-extensions==1.7.1
djangorestframework==3.4.4
